blog/pyproject.toml
[tool.poetry]
name = "fastapi-blog"
version = "0.1.0"
description = "FastAPI blog platform"
authors = ["Sergey Vernigora <volt.awp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.9.1"
aiomcache = "0.8.1"
aioredis = "1.3.1"
aiosignal = "1.3.1"
alembic = "1.12.1"
amqp = "5.2.0"
annotated-types = "0.6.0"
anyio = "3.7.1"
asttokens = "2.4.1"
async-timeout = "4.0.3"
attrs = "23.2.0"
bcrypt = "4.0.1"
billiard = "4.2.0"
cachetools = "5.3.2"
celery = "5.3.6"
cerberus = "1.3.5"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
cryptography = "41.0.5"
decorator = "5.1.1"
distlib = "0.3.7"
dnspython = "2.4.2"
docopt = "0.6.2"
ecdsa = "0.18.0"
email-validator = "2.1.0.post1"
executing = "2.0.1"
fastapi = "0.105.0"
fastapi-cache2 = "0.2.1"
flake8 = "6.1.0"
flower = "2.0.1"
frozenlist = "1.4.1"
greenlet = "3.0.1"
h11 = "0.14.0"
hiredis = "2.3.2"
httpcore = "1.0.2"
httplib2 = "0.22.0"
httptools = "0.6.1"
httpx = "0.25.2"
humanize = "4.9.0"
idna = "3.6"
iniconfig = "2.0.0"
itsdangerous = "2.1.2"
jedi = "0.19.1"
jinja2 = "3.1.2"
kombu = "5.3.4"
mako = "1.3.0"
markupsafe = "2.1.3"
matplotlib-inline = "0.1.6"
mccabe = "0.7.0"
multidict = "6.0.4"
oauthlib = "3.2.2"
orjson = "3.9.10"
packaging = "23.2"
parso = "0.8.3"
passlib = "1.7.4"
pendulum = "3.0.0"
pep517 = "0.13.1"
pexpect = "4.9.0"
pip-api = "0.0.30"
pipreqs = "0.4.13"
platformdirs = "4.1.0"
plette = "0.4.4"
pluggy = "1.3.0"
prometheus-client = "0.19.0"
prompt-toolkit = "3.0.41"
protobuf = "4.25.1"
psycopg2 = "2.9.9"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycodestyle = "2.11.1"
pycparser = "2.21"
pydantic = "2.5.2"
pydantic-extra-types = "2.1.0"
pydantic-settings = "2.1.0"
pydantic-core = "2.14.5"
pyflakes = "3.1.0"
pygments = "2.17.2"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
python-multipart = "0.0.6"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
redis = "5.0.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
requirementslib = "3.0.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.23"
sqlalchemy-utils = "0.41.1"
stack-data = "0.6.3"
starlette = "0.27.0"
time-machine = "2.13.0"
tomlkit = "0.12.3"
tornado = "6.4"
traitlets = "5.14.0"
typing-extensions = "4.8.0"
tzdata = "2023.4"
ujson = "5.8.0"
uritemplate = "4.1.1"
urllib3 = "2.1.0"
uvicorn = "0.24.0.post1"
uvloop = "0.19.0"
vine = "5.1.0"
watchfiles = "0.21.0"
wcwidth = "0.2.12"
websockets = "12.0"
yarg = "0.1.9"
yarl = "1.9.4"
asgiref = "3.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-celery = "0.0.0"
pytest-mock = "3.12.0"
ipython = "8.18.1"
coverage = "7.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
